requirements.txt
cairocffi==1.1.0
cffi==1.14.0
cycler==0.10.0
Cython==0.29.16
fa2==0.3.5
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-igraph==0.8.0
pytz==2019.3
scipy==1.4.1
six==1.14.0
texttable==1.6.2
tqdm==4.44.1
